+++ hvega/tools/viewer.js
@@ -1,0 +1,5 @@
+/*
+ * Can we view the visualizations?
+ */
+
+var embed = require("vege-embed");